+++ tests/webview/lib/repo-selection.test.ts
@@ -12,7 +12,6 @@ beforeAll(() => {
   initializeWebviewConfig({
     autoCenterCommitDetailsView: true,
     dateFormat: "Date & Time",
-    fetchAvatars: false,
     graphColours: [],
     graphStyle: "rounded",
     initialLoadCommits: 300,